+++ custom/ilsgateway/tests/handlers/stockonhand.py
@@ -10,9 +10,6 @@ from custom.ilsgateway.tests.handlers.utils import ILSTestScript
 
 class ILSSoHTest(ILSTestScript):
 
-    def setUp(self):
-        super(ILSSoHTest, self).setUp()
-
     def test_stock_on_hand(self):
         with localize('sw'):
             response = unicode(SOH_CONFIRM)
@@ -95,114 +92,6 @@ class ILSSoHTest(ILSTestScript):
                 this_pkmax = max(this_pkmax, pr.pk)
             pkmax = max(this_pkmax, pkmax)
 
-    def test_product_aliases(self):
-        with localize('sw'):
-            response = unicode(SOH_CONFIRM)
-
-        script = """
-            5551234 > Hmk iucd 400
-            5551234 < {}
-        """.format(response)
-        self.run_script(script)
-
-        script = """
-            5551234 > Hmk Depo 569
-            5551234 < {}
-        """.format(response)
-        self.run_script(script)
-
-        script = """
-            5551234 > Hmk IMPL 678
-            5551234 < %(soh_confirm)s
-        """ % {"soh_confirm": response}
-        self.run_script(script)
-
-    def test_stock_on_hand_delimiter_standard(self):
-        with localize('sw'):
-            response = unicode(SOH_CONFIRM)
-
-        # standard spacing
-        script = """
-            5551234 > hmk fs100 md100 ff100 dx100 bp100 pc100 qi100
-            5551234 < %(soh_confirm)s
-        """ % {"soh_confirm": response}
-        self.run_script(script)
-
-    def test_stock_on_hand_delimiter_no_spaces(self):
-        with localize('sw'):
-            response = unicode(SOH_CONFIRM)
-
-        # no spaces
-        script = """
-            5551234 > hmk fs100md100ff100dx100bp100pc100qi100
-            5551234 < %(soh_confirm)s
-        """ % {"soh_confirm": unicode(response)}
-        self.run_script(script)
-
-    def test_stock_on_hand_delimiters_mixed_spacing(self):
-        with localize('sw'):
-            response = unicode(SOH_CONFIRM)
-        # no spaces
-        script = """
-            5551234 > hmk fs100 md 100 ff100 dx  100bp   100 pc100 qi100
-            5551234 < %(soh_confirm)s
-        """ % {"soh_confirm": unicode(response)}
-        self.run_script(script)
-
-    def test_stock_on_hand_delimiters_all_spaced_out(self):
-        with localize('sw'):
-            response = unicode(SOH_CONFIRM)
-        # all space delimited
-        script = """
-            5551234 > hmk fs 100 md 100 ff 100 dx 100 bp 100 pc 100 qi 100
-            5551234 < %(soh_confirm)s
-        """ % {"soh_confirm": response}
-        self.run_script(script)
-
-    def test_stock_on_hand_delimiters_commas(self):
-        with localize('sw'):
-            response = unicode(SOH_CONFIRM)
-
-        # commas
-        script = """
-            5551234 > hmk fs100,md100,ff100
-            5551234 < %(soh_confirm)s
-        """ % {"soh_confirm": response}
-        self.run_script(script)
-
-    def test_stock_on_hand_delimiters_commas_and_spaces(self):
-        with localize('sw'):
-            response = unicode(SOH_CONFIRM)
-
-        # commas
-        script = """
-            5551234 > hmk fs100, md100, ff100
-            5551234 < %(soh_confirm)s
-        """ % {"soh_confirm": response}
-        self.run_script(script)
-
-    def test_stock_on_hand_delimiters_extra_spaces(self):
-        with localize('sw'):
-            response = unicode(SOH_CONFIRM)
-
-        # extra spaces
-        script = """
-            5551234 > hmk fs  100   md    100     ff      100       pc        100
-            5551234 < %(soh_confirm)s
-        """ % {"soh_confirm": response}
-        self.run_script(script)
-
-    def test_stock_on_hand_mixed_delimiters_and_spacing(self):
-        with localize('sw'):
-            response = unicode(SOH_CONFIRM)
-
-        # mixed - commas, spacing
-        script = """
-            5551234 > hmk fs100 , md100,ff 100 pc  100  qi,       1000,bp, 100, dx,100
-            5551234 < %(soh_confirm)s
-        """ % {"soh_confirm": response}
-        self.run_script(script)
-
     def test_stock_on_hand_invalid_code(self):
         with localize('sw'):
             response = unicode(SOH_BAD_FORMAT)
@@ -216,62 +105,3 @@ class ILSSoHTest(ILSTestScript):
             sql_product__code='ff',
             case_id=self.facility_sp_id
         ).stock_on_hand, 100)
-
-    def test_stock_on_hand_language_swahili(self):
-        with localize('sw'):
-            response = unicode(SOH_CONFIRM)
-        script = """
-            5551234 > hmk fs100md100
-            5551234 < %(soh_confirm)s
-        """ % {"soh_confirm": response}
-        self.run_script(script)
-
-    def test_stock_on_hand_language_english(self):
-        with localize('en'):
-            response = unicode(LANGUAGE_CONFIRM)
-            response2 = unicode(SOH_CONFIRM)
-
-        language_message = """
-            5551234 > language en
-            5551234 < {0}
-        """.format(unicode(response % dict(language='English')))
-        self.run_script(language_message)
-
-        script = """
-            5551234 > hmk fs100md100
-            5551234 < %(soh_confirm)s
-        """ % {"soh_confirm": response2}
-        self.run_script(script)
-
-        with localize('sw'):
-            response = unicode(LANGUAGE_CONFIRM)
-
-        language_message = """
-            5551234 > language sw
-            5551234 < {0}
-        """.format(unicode(response % dict(language='Swahili')))
-        self.run_script(language_message)
-
-    def test_multiline_message(self):
-        quantities = {
-            'fs': 100,
-            'md': 100,
-            'ff': 100,
-            'pc': 100
-        }
-        message = """
-            hmk
-            fs 100 md 100 ff 100 pc 100
-        """
-        verified_number = get_two_way_number_for_recipient(self.user1)
-        msg = incoming(
-            verified_number.phone_number, message, verified_number.backend_id
-        )
-        self.assertIsNotNone(msg)
-
-        stock_states = StockState.objects.filter(
-            case_id=self.facility_sp_id
-        ).values_list('sql_product__code', 'stock_on_hand')
-
-        for product_code, quantity in stock_states:
-            self.assertEqual(quantity, quantities[product_code])
