+++ custom/ilsgateway/tests/handlers/translation.py
@@ -9,21 +9,6 @@ class TranslationTest(ILSTestScript):
     def setUp(self):
         super(TranslationTest, self).setUp()
 
-    def test_soh(self):
-        with localize('sw'):
-            response1 = unicode(SOH_CONFIRM)
-        script = """
-            5551234 > soh jd 400 mc 569
-            5551234 < {0}
-        """.format(response1)
-        self.run_script(script)
-
-        self.run_script(script)
-        self.assertEqual(2, StockState.objects.count())
-        for ps in StockState.objects.all():
-            self.assertEqual(self.user_fac1.location.linked_supply_point().get_id, ps.case_id)
-            self.assertTrue(0 != ps.stock_on_hand)
-
     def test_soh_in_swahili(self):
         self.user_fac1.language = 'en'
         self.user_fac1.save()
